Update SlackSocket.cs
When included in a larger project, there are times when not all assemblies can be loaded. This would cause a loader exception within the static initializer which is likely not the intended outcome. Not sure if there's a better way to check if an assembly can be reflected on other  than a try/catch.